+++ ESMF/GOCART2G_GridComp/NI2G_GridComp/NI2G_GridCompMod.F90
@@ -11,7 +11,7 @@ module NI2G_GridCompMod
 ! !USES:
    use ESMF
    use MAPL
-   use Chem_MieTableMod2G
+   use Chem_Mie2GMod
    use Chem_AeroGeneric
    use iso_c_binding, only: c_loc, c_f_pointer, c_ptr
 
@@ -310,6 +310,9 @@ contains
     real, dimension(4)   :: Vect_Hcts
 !    real, allocatable, dimension(:) :: rmedDU, rmedSS, fnumDU, fnumSS
     integer :: itemCount
+    integer, allocatable, dimension(:)   :: channels_
+    integer                              :: nmom_
+    character(len=:), allocatable        :: file_
 
     __Iam__('Initialize')
 
@@ -481,45 +484,38 @@ contains
     call MAPL_GetResource (MAPL, NUM_BANDS, 'NUM_BANDS:', __RC__)
 
 !   Get file names for the optical tables
-    call ESMF_ConfigGetAttribute (cfg, self%rad_MieTable(instance)%optics_file, &
+    call ESMF_ConfigGetAttribute (cfg, file_ ,&
                                   label="aerosol_radBands_optics_file:", __RC__ )
 
-    allocate (self%rad_MieTable(instance)%channels(NUM_BANDS), __STAT__ )
-    self%rad_MieTable(instance)%nch = NUM_BANDS
+    allocate (channels_(NUM_BANDS), __STAT__ )
 
     call ESMF_ConfigFindLabel(cfg, label="BANDS:", isPresent=bands_are_present, __RC__)
 
     if (bands_are_present) then
-       call ESMF_ConfigGetAttribute (cfg, self%rad_MieTable(instance)%channels, label= "BANDS:", &
-                                    count=self%rad_MieTable(instance)%nch, __RC__)
+       call ESMF_ConfigGetAttribute (cfg, channels_, label= "BANDS:", &
+                                    count=NUM_BANDS, __RC__)
     else
        do i = 1, NUM_BANDS
-          self%rad_MieTable(instance)%channels(i) = i
+          channels_(i) = i
        end do
     endif
-
-    allocate (self%rad_MieTable(instance)%mie_aerosol, __STAT__)
-    self%rad_MieTable(instance)%mie_aerosol = Chem_MieTableCreate (self%rad_MieTable(instance)%optics_file, __RC__)
-    call Chem_MieTableRead (self%rad_MieTable(instance)%mie_aerosol, NUM_BANDS, self%rad_MieTable(instance)%channels, __RC__)
+    self%rad_Mie(instance) = Chem_Mie2G(file_, channels_*1.e-9, __RC__)
+    VERIFY_(status)
+    deallocate(channels_)
 
 !   Create Diagnostics Mie Table
 !   -----------------------------
 !   Get file names for the optical tables
-    call ESMF_ConfigGetAttribute (cfg, self%diag_MieTable(instance)%optics_file, &
+    call ESMF_ConfigGetAttribute (cfg, file_, &
                                   label="aerosol_monochromatic_optics_file:", __RC__ )
-    call ESMF_ConfigGetAttribute (cfg, self%diag_MieTable(instance)%nmom, label="n_moments:", default=0,  __RC__)
-
+    call ESMF_ConfigGetAttribute (cfg, nmom_, label="n_moments:", default=0,  __RC__)
     i = ESMF_ConfigGetLen (universal_cfg, label='aerosol_monochromatic_optics_wavelength_in_nm_from_LUT:', __RC__)
-    self%diag_MieTable(instance)%nch = i
-    allocate (self%diag_MieTable(instance)%channels(self%diag_MieTable(instance)%nch), __STAT__ )
-    call ESMF_ConfigGetAttribute (universal_cfg, self%diag_MieTable(instance)%channels, &
+    allocate (channels_(i), __STAT__ )
+    call ESMF_ConfigGetAttribute (universal_cfg, channels_, &
                                   label= "aerosol_monochromatic_optics_wavelength_in_nm_from_LUT:", __RC__)
-
-    allocate (self%diag_MieTable(instance)%mie_aerosol, __STAT__)
-    self%diag_MieTable(instance)%mie_aerosol = Chem_MieTableCreate (self%diag_MieTable(instance)%optics_file, __RC__ )
-    call Chem_MieTableRead (self%diag_MieTable(instance)%mie_aerosol, self%diag_MieTable(instance)%nch, &
-                            self%diag_MieTable(instance)%channels*1.e-9, rc=status, nmom=self%diag_MieTable(instance)%nmom)
+    self%diag_Mie(instance) = Chem_Mie2G(file_, channels_*1.e-9, nmom=nmom_, __RC__)
     VERIFY_(status)
+    deallocate(channels_)
 
     ! Mie Table instance/index
     call ESMF_AttributeSet(aero, name='mie_table_instance', value=instance, __RC__)
@@ -955,8 +951,8 @@ contains
    allocate(aerosol(ubound(NH4a,1), ubound(NH4a,2), ubound(NH4a,3), 3), __STAT__)
    aerosol(:,:,:,:) = 0.0
    aerosol(:,:,:,1) = NH4a
-   call Aero_Compute_Diags (mie_table=self%diag_MieTable(self%instance), km=self%km, klid=self%klid, nbegin=1, &
-                            nbins=1, channels=self%diag_MieTable(self%instance)%channels*1.0e-9, &
+   call Aero_Compute_Diags (mie=self%diag_Mie(self%instance), km=self%km, klid=self%klid, nbegin=1, &
+                            nbins=1, &
                             wavelengths_profile=self%wavelengths_profile*1.0e-9, &
                             wavelengths_vertint=self%wavelengths_vertint*1.0e-9, &
                             aerosol=aerosol, grav=MAPL_GRAV, tmpu=t, rhoa=airdens, rh=rh2, u=u, v=v, &
@@ -964,8 +960,8 @@ contains
                             sfcmass=NH4SMASS, colmass=NH4CMASS, mass=NH4MASS, conc=NH4CONC, __RC__)
 
    aerosol(:,:,:,1) = NH3
-   call Aero_Compute_Diags (mie_table=self%diag_MieTable(self%instance), km=self%km, klid=self%klid, nbegin=1, &
-                            nbins=1, channels=self%diag_MieTable(self%instance)%channels*1.0e-9, &
+   call Aero_Compute_Diags (mie=self%diag_Mie(self%instance), km=self%km, klid=self%klid, nbegin=1, &
+                            nbins=1, &
                             wavelengths_profile=self%wavelengths_profile*1.0e-9, &
                             wavelengths_vertint=self%wavelengths_vertint*1.0e-9, &
                             aerosol=aerosol, grav=MAPL_GRAV, tmpu=t, rhoa=airdens, rh=rh2, u=u, v=v, &
@@ -973,8 +969,8 @@ contains
                             sfcmass=NH3SMASS, colmass=NH3CMASS, mass=NH3MASS, conc=NH3CONC, __RC__)
 
    aerosol(:,:,:,1) = NO3an1
-   call Aero_Compute_Diags (mie_table=self%diag_MieTable(self%instance), km=self%km, klid=self%klid, nbegin=1, &
-                            nbins=1, channels=self%diag_MieTable(self%instance)%channels*1.0e-9, &
+   call Aero_Compute_Diags (mie=self%diag_Mie(self%instance), km=self%km, klid=self%klid, nbegin=1, &
+                            nbins=1, &
                             wavelengths_profile=self%wavelengths_profile*1.0e-9, &
                             wavelengths_vertint=self%wavelengths_vertint*1.0e-9, &
                             aerosol=aerosol, grav=MAPL_GRAV, tmpu=t, rhoa=airdens, rh=rh2, u=u, v=v, &
@@ -986,8 +982,8 @@ contains
    aerosol(:,:,:,1) = NO3an1
    aerosol(:,:,:,2) = NO3an2
    aerosol(:,:,:,3) = NO3an3
-   call Aero_Compute_Diags (mie_table=self%diag_MieTable(self%instance), km=self%km, klid=self%klid, nbegin=1, &
-                            nbins=3, channels=self%diag_MieTable(self%instance)%channels*1.0e-9, &
+   call Aero_Compute_Diags (mie=self%diag_Mie(self%instance), km=self%km, klid=self%klid, nbegin=1, &
+                            nbins=3,  &
                             wavelengths_profile=self%wavelengths_profile*1.0e-9, &
                             wavelengths_vertint=self%wavelengths_vertint*1.0e-9, &
                             aerosol=aerosol, grav=MAPL_GRAV, tmpu=t, rhoa=airdens, rh=rh2, u=u, v=v, &
@@ -1159,7 +1155,7 @@ contains
     address = transfer(opaque_self, address)
     call c_f_pointer(address, self)
 
-    call mie_ (self%rad_MieTable(instance), nbins, n_bands, offset, q_4d, rh, ext_s, ssa_s, asy_s, __RC__)
+    call mie_ (self%rad_Mie(instance), nbins, n_bands, offset, q_4d, rh, ext_s, ssa_s, asy_s, __RC__)
 
     call ESMF_AttributeGet(state, name='extinction_in_air_due_to_ambient_aerosol', value=fld_name, __RC__)
     if (fld_name /= '') then
@@ -1187,11 +1183,11 @@ contains
   contains
 
 !    subroutine mie_(mie_table, aerosol_names, nb, offset, q, rh, bext_s, bssa_s, basym_s, rc)
-    subroutine mie_(mie_table, nbins, nb, offset, q, rh, bext_s, bssa_s, basym_s, rc)
+    subroutine mie_(mie, nbins, nb, offset, q, rh, bext_s, bssa_s, basym_s, rc)
 
     implicit none
 
-    type(Chem_Mie),                intent(inout) :: mie_table        ! mie table
+    type(Chem_Mie2G) ,             intent(inout) :: mie              ! mie table
     integer,                       intent(in   ) :: nbins            ! number of bins
     integer,                       intent(in )   :: nb               ! number of bands
     integer,                       intent(in )   :: offset           ! bands offset 
@@ -1214,7 +1210,8 @@ contains
      basym_s = 0.0d0
 
     do l = 1, nbins
-       call Chem_MieQuery(mie_table, l, real(offset+1.), q(:,:,:,l), rh, tau=bext, gasym=gasym, ssa=bssa)
+       !tau is converted to bext
+       call mie%Query(l, offset+1., q(:,:,:,l), rh, tau=bext, gasym=gasym, ssa=bssa)
 
        bext_s  = bext_s  +             bext     ! extinction
        bssa_s  = bssa_s  +       (bssa*bext)    ! scattering extinction
@@ -1277,21 +1274,6 @@ contains
 !   --------------
     call ESMF_AttributeGet(state, name='wavelength_for_aerosol_optics', value=wavelength, __RC__)
 
-!   Get wavelength index for Mie Table
-!   ----------------------------------
-!   Channel values are 4.7e-7 5.5e-7 6.7e-7 8.7e-7 [meter]. Their indices are 1,2,3,4 respectively.
-    if ((wavelength .ge. 4.69e-7) .and. (wavelength .le. 4.71e-7)) then
-       mieTable_index = 1.
-    else if ((wavelength .ge. 5.49e-7) .and. (wavelength .le. 5.51e-7)) then
-       mieTable_index = 2.
-    else if ((wavelength .ge. 6.69e-7) .and. (wavelength .le. 6.71e-7)) then
-       mieTable_index = 3.
-    else if ((wavelength .ge. 8.68e-7) .and. (wavelength .le. 8.71e-7)) then
-       mieTable_index = 4.
-    else
-       print*,trim(Iam),' : wavelength of ',wavelength,' is an invalid value.'
-       return
-    end if
 
 !   Pressure at layer edges 
 !   ------------------------
@@ -1337,15 +1319,14 @@ contains
     address = transfer(opaque_self, address)
     call c_f_pointer(address, self)
 
+!   Get wavelength index for Mie Table
+!   ----------------------------------
+!   Channel values are 4.7e-7 5.5e-7 6.7e-7 8.7e-7 [meter]. Their indices are 1,2,3,4 respectively.
+    mieTable_index = self%diag_Mie(instance)%get_index(wavelength, __RC__)
+
     do n = 1, nbins
-      do i = 1, i2
-        do j = 1, j2
-          do k = 1, km
-            call Chem_MieQuery(self%diag_MieTable(instance), n, mieTable_index, q_4d(i,j,k,n), rh(i,j,k), tau=tau(i,j,k), __RC__)
-            tau_s(i,j,k) = tau_s(i,j,k) + tau(i,j,k)
-          end do
-        end do
-      end do
+       call self%diag_Mie(instance)%Query(n, mieTable_index, q_4d(:,:,:,n), rh, tau=tau, __RC__)
+       tau_s = tau_s + tau
     end do
 
     call ESMF_AttributeGet(state, name='monochromatic_extinction_in_air_due_to_ambient_aerosol', value=fld_name, __RC__)