+++ Process_Library/Chem_Mie2GMod.F90
@@ -73,7 +73,6 @@ module Chem_Mie2GMod
       generic   :: Query => Query_0d, Query_1d, Query_2d, Query_3d
       !generic   :: Query => QueryScalar, QueryVector
       procedure :: get_index
-      procedure :: get_wavelengths
    end type Chem_Mie2G
 
    interface Chem_Mie2G
@@ -1031,14 +1030,5 @@ include "MieQuery_xd.H"
 
   end function get_Index
  
-  function get_wavelengths(this) result(wavelengths)
-    real, allocatable :: wavelengths(:)
-    class (Chem_Mie2G), intent(in) :: this
-
-    if (.not. allocated(this%wavelengths)) return
-
-    wavelengths = this%wavelengths
-  end function get_wavelengths
-
 end module Chem_Mie2GMod