+++ Process_Library/GOCART2G_Process.F90
@@ -17,7 +17,7 @@
 
 ! !USES:
 !  Only instrinsic fortran types and functions are allowed.
-   use Chem_MieTableMod2G
+   use Chem_Mie2GMod
    use, intrinsic :: iso_fortran_env, only: IOSTAT_END
 
    implicit none
@@ -3251,7 +3251,7 @@ CONTAINS
 ! !IROUTINE:  Aero_Compute_Diags - Calculate aerosol diagnostics
 !
 ! !INTERFACE:
-   subroutine Aero_Compute_Diags (mie_table, km, klid, nbegin, nbins, rlow, rup, channels, &
+   subroutine Aero_Compute_Diags (mie, km, klid, nbegin, nbins, rlow, rup, &
                                   wavelengths_profile, wavelengths_vertint, aerosol, &
                                   grav, tmpu, rhoa, rh, u, v, delp, ple,tropp, &
                                   sfcmass, colmass, mass, exttau, stexttau, scatau, stscatau,&
@@ -3264,12 +3264,11 @@ CONTAINS
   implicit NONE
 
 ! !INPUT PARAMETERS:
-   type(Chem_Mie),  intent(in) :: mie_table        ! mie table
+   type(Chem_Mie2G),  intent(in) :: mie        ! mie table
    integer, intent(in) :: km, nbegin, nbins
    integer,    intent(in)    :: klid   ! index for pressure lid
    real, optional, dimension(:), intent(in)    :: rlow   ! bin radii - low bounds
    real, optional, dimension(:), intent(in)    :: rup    ! bin radii - upper bounds
-   real, dimension(:), intent(in)    :: channels
    real, dimension(:), intent(in)    :: wavelengths_profile
    real, dimension(:), intent(in)    :: wavelengths_vertint
    real, dimension(:,:,:,:), intent(in) :: aerosol     !
@@ -3322,11 +3321,11 @@ CONTAINS
 
 ! !Local Variables
    character(len=*), parameter :: myname = 'Aero_Compute_Diags'
-   integer :: i, j, k, n, w, ios, nch
+   integer :: i, j, k, n, w, ios, status
    integer :: i1 =1, i2, j1=1, j2
    real :: ilam550, ilam470, ilam870
    real, allocatable, dimension(:) :: wavelengths_index_profile, wavelengths_index_vertint
-   real :: tau, ssa
+   real, allocatable, dimension(:,:,:) :: tau, ssa
 !   real :: fPMfm(nbins)  ! fraction of bin with particles diameter < 1.0 um
 !   real :: fPM25(nbins)  ! fraction of bin with particles diameter < 2.5 um
    real, dimension(:), allocatable :: fPMfm  ! fraction of bin with particles diameter < 1.0 um
@@ -3347,7 +3346,6 @@ CONTAINS
 
 !  Initialize local variables
 !  --------------------------
-   nch = size(channels)
    i2 = size(rhoa,1)
    j2 = size(rhoa,2)
    allocate(fPMfm(nbins))
@@ -3360,50 +3358,24 @@ CONTAINS
    allocate(wavelengths_index_vertint(size(wavelengths_vertint)))
    wavelengths_index_profile = 0.
    wavelengths_index_vertint = 0.
-   ilam550 = 1.
-   ilam470 = 0.
-   ilam870 = 0.
-   if(nch .gt. 1) then
-      do i = 1, nch
-         if ( channels(i) .ge. 5.49e-7 .and. &
-              channels(i) .le. 5.51e-7) ilam550 = i
-         if ( channels(i) .ge. 4.69e-7 .and. &
-              channels(i) .le. 4.71e-7) ilam470 = i
-         if ( channels(i) .ge. 8.69e-7 .and. &
-              channels(i) .le. 8.71e-7) ilam870 = i
-      enddo
-   endif
+
+   ilam470 = mie%get_index(4.70e-7)
+   if(ilam470 <= 0) ilam470 = 0.
+
+   ilam550 = mie%get_index(5.50e-7)
+   if(ilam550 <= 0) ilam550 = 1.
+
+   ilam870 = mie%get_index(8.70e-7)
+   if(ilam870 <= 0) ilam870 = 0.
 
    ! Channel values are 4.7e-7 5.5e-7 6.7e-7 8.7e-7 [meter]. Their indices are 1,2,3,4 respectively.
    do i = 1, size(wavelengths_profile)
-      if ((wavelengths_profile(i) .ge. 5.49e-7) .and. (wavelengths_profile(i) .le. 5.51e-7)) then
-         wavelengths_index_profile(i) = 2.
-      else if ((wavelengths_profile(i) .ge. 4.69e-7) .and. (wavelengths_profile(i) .le. 4.71e-7)) then
-         wavelengths_index_profile(i) = 1.
-      else if ((wavelengths_profile(i) .ge. 6.69e-7) .and. (wavelengths_profile(i) .le. 6.71e-7)) then
-         wavelengths_index_profile(i) = 3.
-      else if ((wavelengths_profile(i) .ge. 8.68e-7) .and. (wavelengths_profile(i) .le. 8.71e-7)) then
-         wavelengths_index_profile(i) = 4.
-      else
-         print*,'wavelengths_profile of ',wavelengths_profile(i),' is an invalid value.'
-         return
-      end if
+      wavelengths_index_profile(i) = mie%get_index(wavelengths_profile(i), __RC__)
    end do
 
    ! Channel values are 4.7e-7 5.5e-7 6.7e-7 8.7e-7 [meter]. Their indices are 1,2,3,4 respectively.
    do i = 1, size(wavelengths_vertint)
-      if ((wavelengths_vertint(i) .ge. 5.49e-7) .and. (wavelengths_vertint(i) .le. 5.51e-7)) then
-         wavelengths_index_vertint(i) = 2.
-      else if ((wavelengths_vertint(i) .ge. 4.69e-7) .and. (wavelengths_vertint(i) .le. 4.71e-7)) then
-         wavelengths_index_vertint(i) = 1.
-      else if ((wavelengths_vertint(i) .ge. 6.69e-7) .and. (wavelengths_vertint(i) .le. 6.71e-7)) then
-         wavelengths_index_vertint(i) = 3.
-      else if ((wavelengths_vertint(i) .ge. 8.68e-7) .and. (wavelengths_vertint(i) .le. 8.71e-7)) then
-         wavelengths_index_vertint(i) = 4.
-      else
-         print*,'wavelengths_vertint of ',wavelengths_vertint(i),' is an invalid value.'
-         return
-      end if
+      wavelengths_index_vertint(i) = mie%get_index(wavelengths_vertint(i), __RC__)
    end do
 
 !  Determine if going to do Angstrom parameter calculation
@@ -3522,6 +3494,8 @@ CONTAINS
       end do
    endif
 
+   allocate(tau(i1:i2,j1:j2,km),source = 0.)
+   allocate(ssa(i1:i2,j1:j2,km),source = 0.)
 !  Calculate the extinction and/or scattering AOD
    if( present(extcoef)  .or. &
        present(scacoef) ) then
@@ -3530,27 +3504,19 @@ CONTAINS
       if( present(scacoef) ) scacoef = 0.
 
       do n = nbegin, nbins
-       do w = 1, size(wavelengths_profile)
-         do k = klid, km
-            do j = j1, j2
-               do i = i1, i2
-!                call Chem_MieQuery(mie_table, n, ilam550, &
-                 call Chem_MieQuery(mie_table, n, wavelengths_index_profile(w), &
-                 aerosol(i,j,k,n)*delp(i,j,k)/grav, &
-                 rh(i,j,k), tau=tau, ssa=ssa)
-
-!                Calculate the total ext. and scat. coefficients
-                 if( present(extcoef) ) then
-                     extcoef(i,j,k,w) = extcoef(i,j,k,w) + &
-                                      tau * (grav * rhoa(i,j,k) / delp(i,j,k))
-                 endif
-                 if( present(scacoef) ) then
-                    scacoef(i,j,k,w) = scacoef(i,j,k,w) + &
-                                     ssa * tau * (grav * rhoa(i,j,k) / delp(i,j,k))
-                 endif
-               enddo !i
-            enddo !j
-         enddo !k
+        do w = 1, size(wavelengths_profile)
+          call mie%Query(n, wavelengths_index_profile(w), &
+                 aerosol(:,:,:,n)*delp/grav, &
+                 rh, tau=tau, ssa=ssa)
+!         Calculate the total ext. and scat. coefficients
+          if ( present(extcoef) ) then
+             extcoef(:,:,:,w) = extcoef(:,:,:,w) + &
+                               tau * (grav * rhoa / delp)
+          endif
+          if ( present(scacoef) ) then
+             scacoef(:,:,:,w) = scacoef(:,:,:,w) + &
+                               ssa * tau * (grav * rhoa / delp)
+          endif
        enddo !wavelengths_profile
       enddo !nbins
     end if !present(extcoef)...
@@ -3575,72 +3541,63 @@ CONTAINS
       if( present(exttaufm) ) exttaufm = 0.
       if( present(scataufm) ) scataufm = 0.
 
-       do w = 1, size(wavelengths_vertint)
-      do n = nbegin, nbins
-!       do w = 1, size(wavelengths_vertint)
-         do k = klid, km
-            do j = j1, j2
-               do i = i1, i2
+      do w = 1, size(wavelengths_vertint)
+        do n = nbegin, nbins
+           call mie%Query(n, wavelengths_index_vertint(w), &
+                aerosol(:,:,:,n)*delp/grav, &
+                rh, tau=tau, ssa=ssa)
+           do k = klid, km
+!             Integrate in the vertical
+              if( present(exttau) ) exttau(:,:,w) = exttau(:,:,w) + tau(:,:,k)
+              if( present(stexttau) ) then
+                 where (ple(:,:,k) .le. tropp) 
+                    stexttau(:,:,w) = stexttau(:,:,w) + tau(:,:,k)
+                 elsewhere(ple(:,:,k) .gt. tropp .and. ple(:,:,k-1) .lt. tropp) 
+                    stexttau(:,:,w) = stexttau(:,:,w) + log(tropp/ple(:,:,k-1))/log(ple(:,:,k)/ple(:,:,k-1))*tau(:,:,k)
+                 endwhere
+              endif
 
-!                call Chem_MieQuery(mie_table, n, ilam550, &
-                 call Chem_MieQuery(mie_table, n, wavelengths_index_vertint(w), &
-                 aerosol(i,j,k,n)*delp(i,j,k)/grav, &
-                 rh(i,j,k), tau=tau, ssa=ssa)
-
-!                Integrate in the vertical
-                 if( present(exttau) ) exttau(i,j,w) = exttau(i,j,w) + tau
-                 if( present(stexttau) ) then
-                    if (ple(i,j,k) .le. tropp(i,j)) then
-                        stexttau(i,j,w) = stexttau(i,j,w) + tau
-                    elseif(ple(i,j,k) .gt. tropp(i,j) .and. ple(i,j,k-1) .lt. tropp(i,j)) then
-                        stexttau(i,j,w) = stexttau(i,j,w) + log(tropp(i,j)/ple(i,j,k-1))/log(ple(i,j,k)/ple(i,j,k-1))*tau
-                    endif
-                 endif
-
-                 if( present(exttaufm) ) then
-                    if( NO3nFlag_ ) then
-                       exttaufm(i,j,w) = exttaufm(i,j,w) + tau
-                    else
-                       exttaufm(i,j,w) = exttaufm(i,j,w) + tau * fPMfm(n)
-                    end if
+              if( present(exttaufm) ) then
+                 if( NO3nFlag_ ) then
+                    exttaufm(:,:,w) = exttaufm(:,:,w) + tau(:,:,k)
+                 else
+                    exttaufm(:,:,w) = exttaufm(:,:,w) + tau(:,:,k) * fPMfm(n)
                  end if
+              end if
 
-                 if( present(exttau25) ) then
-                    if( NO3nFlag_ ) then
-                       exttau25(i,j,w) = exttau25(i,j,w) + tau
-                    else
-                       exttau25(i,j,w) = exttau25(i,j,w) + tau * fPM25(n)
-                    end if
+              if( present(exttau25) ) then
+                 if( NO3nFlag_ ) then
+                    exttau25(:,:,w) = exttau25(:,:,w) + tau(:,:,k)
+                 else
+                    exttau25(:,:,w) = exttau25(:,:,w) + tau(:,:,k) * fPM25(n)
                  end if
+              end if
 
-                 if( present(scatau) ) scatau(i,j,w) = scatau(i,j,w) + tau*ssa
-                 if( present(stscatau) ) then
-                    if (ple(i,j,k) .le. tropp(i,j)) then
-                        stscatau(i,j,w) = stscatau(i,j,w) + tau*ssa
-                    elseif(ple(i,j,k) .gt. tropp(i,j) .and. ple(i,j,k-1) .lt. tropp(i,j)) then
-                        stscatau(i,j,w) = stscatau(i,j,w) + log(tropp(i,j)/ple(i,j,k-1))/log(ple(i,j,k)/ple(i,j,k-1))*tau*ssa
-                    endif
-                 endif
-                 if( present(scataufm) ) then
-                    if( NO3nFlag_ ) then
-                       scataufm(i,j,w) = scataufm(i,j,w) + tau * ssa
-                    else
-                       scataufm(i,j,w) = scataufm(i,j,w) + tau * ssa * fPMfm(n)
-                    end if
+              if( present(scatau) ) scatau(:,:,w) = scatau(:,:,w) + tau(:,:,k)*ssa(:,:,k)
+              if( present(stscatau) ) then
+                 where (ple(:,:,k) .le. tropp) 
+                    stscatau(:,:,w) = stscatau(:,:,w) + tau(:,:,k)*ssa(:,:,k)
+                 elsewhere(ple(:,:,k) .gt. tropp .and. ple(:,:,k-1) .lt. tropp) 
+                    stscatau(:,:,w) = stscatau(:,:,w) + log(tropp/ple(:,:,k-1))/log(ple(:,:,k)/ple(:,:,k-1))*tau(:,:,k)*ssa(:,:,k)
+                 endwhere
+              endif
+              if( present(scataufm) ) then
+                 if( NO3nFlag_ ) then
+                    scataufm(:,:,w) = scataufm(:,:,w) + tau(:,:,k) * ssa(:,:,k)
+                 else
+                    scataufm(:,:,w) = scataufm(:,:,w) + tau(:,:,k) * ssa(:,:,k) * fPMfm(n)
                  end if
+              end if
 
-                 if( present(scatau25) ) then
-                    if( NO3nFlag_ ) then
-                       scatau25(i,j,w) = scatau25(i,j,w) + tau * ssa
-                    else
-                       scatau25(i,j,w) = scatau25(i,j,w) + tau * ssa * fPM25(n)
-                    end if
+              if( present(scatau25) ) then
+                 if( NO3nFlag_ ) then
+                    scatau25(:,:,w) = scatau25(:,:,w) + tau(:,:,k) * ssa(:,:,k)
+                 else
+                    scatau25(:,:,w) = scatau25(:,:,w) + tau(:,:,k) * ssa(:,:,k) * fPM25(n)
                  end if
-
-               enddo !i
-            enddo !j
-         enddo !k
-       enddo !wavelengths_vertint
+              end if
+           enddo !k
+        enddo !wavelengths_vertint
       enddo !nbins
    endif !present(exttau)...
 
@@ -3654,30 +3611,27 @@ CONTAINS
 
       do n = nbegin, nbins
 
-!      Select the name for species
-       do k = klid, km
-        do j = j1, j2
-         do i = i1, i2
-          call Chem_MieQuery(mie_table, n, ilam470, &
-              aerosol(i,j,k,n)*delp(i,j,k)/grav, &
-              rh(i,j,k), tau=tau)
-          tau470(i,j) = tau470(i,j) + tau
-
-          call Chem_MieQuery(mie_table, n, ilam870, &
-              aerosol(i,j,k,n)*delp(i,j,k)/grav, &
-              rh(i,j,k), tau=tau)
-          tau870(i,j) = tau870(i,j) + tau
-         enddo
+!       Select the name for species
+        call mie%Query(n, ilam470, &
+              aerosol(:,:,:,n)*delp/grav, &
+              rh, tau=tau)
+        do k = klid, km
+          tau470 = tau470 + tau(:,:,k)
         enddo
-       enddo
 
+        call mie%Query(n, ilam870, &
+              aerosol(:,:,:,n)*delp/grav, &
+              rh, tau=tau)
+        do k = klid, km
+          tau870 = tau870 + tau(:,:,k)
+        enddo
       enddo  ! nbins
 
       angstrom(i1:i2,j1:j2) = &
         -log(tau470(i1:i2,j1:j2)/tau870(i1:i2,j1:j2)) / &
          log(470./870.)
    endif
-
+   deallocate(tau,ssa)
    __RETURN__(__SUCCESS__)
    end subroutine Aero_Compute_Diags
 !====================================================================
@@ -4272,7 +4226,7 @@ CONTAINS
 ! !INTERFACE:
 !
 
-   subroutine CAEmission (mie_table, km, nbins, cdt, grav, prefix, ratPOM, eAircraftfuel, aircraft_fuel_src, &
+   subroutine CAEmission (mie, km, nbins, cdt, grav, prefix, ratPOM, eAircraftfuel, aircraft_fuel_src, &
                            aviation_lto_src, aviation_cds_src, aviation_crs_src, &
                            fHydrophobic, pblh, tmpu, rhoa, rh, aerosolPhilic, aerosolPhobic, &
                            delp, aviation_layers, &
@@ -4284,7 +4238,7 @@ CONTAINS
   implicit NONE
 
 ! !INPUT PARAMETERS:
-   type(Chem_Mie),  intent(in) :: mie_table        ! mie table
+   type(Chem_Mie2G),  intent(in) :: mie        ! mie table
    integer, intent(in) :: km     ! total model levels
    integer, intent(in) :: nbins  ! number of aerosol size bins
    real, intent(in)    :: cdt    ! chemistry model time-step [sec]
@@ -4355,11 +4309,12 @@ CONTAINS
 
    real, dimension(:,:), allocatable          :: f_bb_        ! scaling factor for BB emissions based on maximum allowed exttau
    real, dimension(:,:), allocatable          :: exttau_bb_   ! increment of exttau due to BB during the current time step
-   real, allocatable, dimension(:,:,:,:) :: qa_bb_       ! increment of qa due to BB during the current time step (nbins,i1:i2,j1:j2:km)
+   real, allocatable, dimension(:,:,:,:) :: qa_bb_       ! increment of qa due to BB during the current time step (nbins,i1:i2,j1:j2:km) 
+                                                         ! W.Jiang note, changed to (i1:i2,j1:j2,km,nbins) for efficiency
    real                                  :: cutoff_bb_exttau
-   integer                               :: nch, idx
+   integer                               :: idx
    real                                  :: ilam550
-   real                                  :: tau, ssa
+   real, dimension(:,:,:), allocatable   :: tau
    character(len=255)                    :: qname
    real, parameter                       :: max_bb_exttau = 30.0
 
@@ -4368,7 +4323,6 @@ CONTAINS
 
 !  Source function terms for SOA from Anthropogenic VOCs
    real :: srcSOAanthro = 0.0
-
 !  Initialize local variables
 !  --------------------------
    i2 = size(rhoa,1)
@@ -4467,7 +4421,7 @@ CONTAINS
 
 !   Limit biomass burning emissions
 !   -------------------------------
-    allocate(qa_bb_(nbins,i1:i2,j1:j2,km))
+    allocate(qa_bb_(i1:i2,j1:j2,km,nbins))
     qa_bb_ = 0.0
 
     p0 = ps
@@ -4508,44 +4462,29 @@ K_LOOP_BB: do k = km, 1, -1
 !   -----------------------------------
     factor = cdt * grav / delp(:,:,k)
 
-    qa_bb_(1,:,:,k) = factor * srcHydrophobic
-    qa_bb_(2,:,:,k) = factor * srcHydrophilic
+    qa_bb_(:,:,k,1) = factor * srcHydrophobic
+    qa_bb_(:,:,k,2) = factor * srcHydrophilic
 
    end do K_LOOP_BB
-
-    nch   = mie_table%nch
 
 !   Get the wavelength indices
 !   --------------------------
 !   Must provide ilam550 for AOT calculation
-    ilam550 = 1.
-    if(nch .gt. 1) then
-     do i = 1, nch
-      if ( mie_table%channels(i) .ge. 5.49e-7 .and. &
-           mie_table%channels(i) .le. 5.51e-7) ilam550 = i
-     enddo
-    endif
-
+    ilam550 = mie%get_index(5.50e-7)
+    if (ilam550 <=0) ilam550 = 1.
 !  Calculate the extinction and/or scattering AOD
 
    exttau_bb_(i1:i2,j1:j2) = 0.0
-
+   allocate(tau(i1:i2,j1:j2,km), source = 0.)
    do n = 1, nbins
 !     Select the name for species and the index
-      do k = 1, km
-       do j = j1, j2
-        do i = i1, i2
-         call Chem_MieQuery(mie_table, n, ilam550, &
-              qa_bb_(n,i,j,k)*delp(i,j,k)/grav, &
-              rh(i,j,k), tau=tau, ssa=ssa)
-
+     call mie%Query(n, ilam550, &
+              qa_bb_(:,:,:,n)*delp(:,:,:)/grav, &
+              rh, tau=tau)
+     do k = 1, km
 !        Integrate in the vertical
-         exttau_bb_(i,j) = exttau_bb_(i,j) + tau
-
-        enddo
-       enddo
-      enddo
-
+        exttau_bb_(:,:) = exttau_bb_(:,:) + tau(:,:,k)
+     enddo
    enddo  ! nbins
 
    f_bb_ = 1.0
@@ -4559,7 +4498,7 @@ K_LOOP_BB: do k = km, 1, -1
     enddo
    enddo
 
-   deallocate(qa_bb_)
+   deallocate(qa_bb_, tau)
 
 !  Now update the tracer mixing ratios with the aerosol sources
 !  ------------------------------------------------------------
@@ -6753,7 +6692,7 @@ K_LOOP: do k = km, 1, -1
 !BOP
 ! !IROUTINE: SU_Compute_Diags
 
-   subroutine SU_Compute_Diags ( km, klid, rmed, sigma, rhop, grav, pi, nSO4, mie_table, channels, &
+   subroutine SU_Compute_Diags ( km, klid, rmed, sigma, rhop, grav, pi, nSO4, mie, &
                                  wavelengths_profile, wavelengths_vertint, &
                                  tmpu, rhoa, delp, ple, tropp,rh, u, v, &
                                  DMS, SO2, SO4, MSA, &
@@ -6776,8 +6715,7 @@ K_LOOP: do k = km, 1, -1
    real, intent(in)    :: grav  ! gravity [m/sec]
    real, intent(in)    :: pi    ! pi constant
    integer, intent(in) :: nSO4  ! index of SO4 relative to other internal variables
-   type(Chem_Mie), intent(in) :: mie_table   ! mie table
-   real, dimension(:), intent(in)  :: channels
+   type(Chem_Mie2G), intent(in) :: mie   ! mie table
    real, dimension(:), intent(in)  :: wavelengths_profile
    real, dimension(:), intent(in)  :: wavelengths_vertint
    real, pointer, dimension(:,:,:), intent(in) :: tmpu    ! temperature [K]
@@ -6828,9 +6766,9 @@ K_LOOP: do k = km, 1, -1
 !  29july2020, E.Sherman - refactored for process library
 
 ! !Local Variables
-   integer :: i, j, k, w, i1=1, j1=1, i2, j2, nch
+   integer :: i, j, k, w, i1=1, j1=1, i2, j2, status
    real, allocatable, dimension(:)  :: wavelengths_index_profile, wavelengths_index_vertint
-   real :: tau, ssa
+   real, dimension(:,:,:), allocatable :: tau, ssa
    real, dimension(:,:), allocatable :: tau470, tau870
    real    :: ilam550, ilam470, ilam870
    logical :: do_angstrom
@@ -6840,8 +6778,6 @@ K_LOOP: do k = km, 1, -1
 !EOP
 !-------------------------------------------------------------------------
 !  Begin
-
-   nch = size(channels)
    j2 = ubound(tmpu, 2)
    i2 = ubound(tmpu, 1)
 
@@ -6850,19 +6786,16 @@ K_LOOP: do k = km, 1, -1
 !  Get the wavelength indices
 !  --------------------------
 !  Must provide ilam550 for AOT calculation
-   ilam550 = 1.
-   ilam470 = 0.
-   ilam870 = 0.
-   if(nch .gt. 1) then
-      do i = 1, nch
-         if ( channels(i) .ge. 5.49e-7 .and. &
-              channels(i) .le. 5.51e-7) ilam550 = i
-         if ( channels(i) .ge. 4.69e-7 .and. &
-              channels(i) .le. 4.71e-7) ilam470 = i
-         if ( channels(i) .ge. 8.69e-7 .and. &
-              channels(i) .le. 8.71e-7) ilam870 = i
-      enddo
-   endif
+
+   ilam470 = mie%get_index(4.70e-7)
+   if(ilam470 <= 0) ilam470 = 0.
+
+   ilam550 = mie%get_index(5.50e-7)
+   if(ilam550 <= 0) ilam550 = 1.
+
+   ilam870 = mie%get_index(8.70e-7)
+   if(ilam870 <= 0) ilam870 = 0.
+
 
    allocate(wavelengths_index_profile(size(wavelengths_profile)))
    allocate(wavelengths_index_vertint(size(wavelengths_vertint)))
@@ -6871,34 +6804,14 @@ K_LOOP: do k = km, 1, -1
 
    ! Channel values are 4.7e-7 5.5e-7 6.7e-7 8.7e-7 [meter]. Their indices are 1,2,3,4 respectively.
    do i = 1, size(wavelengths_profile)
-      if ((wavelengths_profile(i) .ge. 5.49e-7) .and. (wavelengths_profile(i) .le. 5.51e-7)) then
-         wavelengths_index_profile(i) = 2.
-      else if ((wavelengths_profile(i) .ge. 4.69e-7) .and. (wavelengths_profile(i) .le. 4.71e-7)) then
-         wavelengths_index_profile(i) = 1.
-      else if ((wavelengths_profile(i) .ge. 6.69e-7) .and. (wavelengths_profile(i) .le. 6.71e-7)) then
-         wavelengths_index_profile(i) = 3.
-      else if ((wavelengths_profile(i) .ge. 8.69e-7) .and. (wavelengths_profile(i) .le. 8.71e-7)) then
-         wavelengths_index_profile(i) = 4.
-      else
-         print*,'wavelengths_profile of ',wavelengths_profile(i),' is an invalid value.'
-         return
-      end if
+      ! WJiang note: is this change correct??
+      wavelengths_index_profile(i) = mie%get_index(wavelengths_profile(i), __RC__)
    end do
 
    ! Channel values are 4.7e-7 5.5e-7 6.7e-7 8.7e-7 [meter]. Their indices are 1,2,3,4 respectively.
    do i = 1, size(wavelengths_vertint)
-      if ((wavelengths_vertint(i) .ge. 5.49e-7) .and. (wavelengths_vertint(i) .le. 5.51e-7)) then
-         wavelengths_index_vertint(i) = 2.
-      else if ((wavelengths_vertint(i) .ge. 4.69e-7) .and. (wavelengths_vertint(i) .le. 4.71e-7)) then
-         wavelengths_index_vertint(i) = 1.
-      else if ((wavelengths_vertint(i) .ge. 6.69e-7) .and. (wavelengths_vertint(i) .le. 6.71e-7)) then
-         wavelengths_index_vertint(i) = 3.
-      else if ((wavelengths_vertint(i) .ge. 8.69e-7) .and. (wavelengths_vertint(i) .le. 8.71e-7)) then
-         wavelengths_index_vertint(i) = 4.
-      else
-         print*,'wavelengths_profile of ',wavelengths_profile(i),' is an invalid value.'
-         return
-      end if
+      ! WJiang note: is this change correct??
+      wavelengths_index_vertint(i) = mie%get_index(wavelengths_vertint(i), __RC__)
    end do
 
 !  Determine if going to do Angstrom parameter calculation
@@ -7008,32 +6921,27 @@ K_LOOP: do k = km, 1, -1
    endif
 
 !  Calculate the extinction and/or scattering AOD
+   allocate(tau(i1:i2,j1:j2,km), source = 0.)
+   allocate(ssa(i1:i2,j1:j2,km), source = 0.)
    if( associated(extcoef) .or. associated(scacoef) ) then
 
       if (associated(extcoef)) extcoef = 0.
       if (associated(scacoef)) scacoef = 0.
 
       do w = 1, size(wavelengths_profile)
-       do k = klid, km
-        do j = j1, j2
-         do i = i1, i2
-          call Chem_MieQuery(mie_table, 1, wavelengths_index_profile(w), & ! Only SO4 exists in the MieTable, so its index is 1
-              SO4(i,j,k)*delp(i,j,k)/grav, &
-              rh(i,j,k), tau=tau, ssa=ssa)
+         call mie%Query(1, wavelengths_index_profile(w), & ! Only SO4 exists in the MieTable, so its index is 1
+              SO4*delp/grav, &
+              rh, tau=tau, ssa=ssa)
 
 !         Calculate the total ext. and scat. coefficients
-          if( associated(extcoef) ) then
-              extcoef(i,j,k,w) = extcoef(i,j,k,w) + &
-                              tau * (grav * rhoa(i,j,k) / delp(i,j,k))
-          endif
-          if( associated(scacoef) ) then
-              scacoef(i,j,k,w) = scacoef(i,j,k,w) + &
-                              ssa * tau * (grav * rhoa(i,j,k) / delp(i,j,k))
-          endif
-
-         enddo
-        enddo
-       enddo
+         if( associated(extcoef) ) then
+              extcoef(:,:,:,w) = extcoef(:,:,:,w) + &
+                              tau * (grav * rhoa / delp)
+         endif
+         if( associated(scacoef) ) then
+              scacoef(:,:,:,w) = scacoef(:,:,:,w) + &
+                              ssa * tau * (grav * rhoa / delp)
+         endif
       enddo
    endif
 
@@ -7046,42 +6954,36 @@ K_LOOP: do k = km, 1, -1
       if (associated(stscatau)) stscatau = 0.
 
       do w = 1, size(wavelengths_vertint)
-       do k = klid, km
-        do j = j1, j2
-         do i = i1, i2
-            call Chem_MieQuery(mie_table, 1, wavelengths_index_vertint(w), & ! Only SO4 exists in the MieTable, so its index is 1
-            SO4(i,j,k)*delp(i,j,k)/grav, &
-            rh(i,j,k), tau=tau, ssa=ssa)
+         call mie%Query(1, wavelengths_index_vertint(w), & ! Only SO4 exists in the MieTable, so its index is 1
+            SO4*delp/grav, &
+            rh, tau=tau, ssa=ssa)
 
-!         Integrate in the vertical
-          if( associated(exttau) ) then
-           exttau(i,j,w) = exttau(i,j,w) + tau
-          endif
+         do k = klid, km
+!           Integrate in the vertical
+            if ( associated(exttau) ) then
+               exttau(:,:,w) = exttau(:,:,w) + tau(:,:,k)
+            endif
 
-          if(associated(stexttau) ) then
-             if (ple(i,j,k) .le. tropp(i,j)) then
-                 stexttau(i,j,w) = stexttau(i,j,w) + tau
-             elseif(ple(i,j,k) .gt. tropp(i,j) .and. ple(i,j,k-1) .lt. tropp(i,j)) then
-                 stexttau(i,j,w) = stexttau(i,j,w) + log(tropp(i,j)/ple(i,j,k-1))/log(ple(i,j,k)/ple(i,j,k-1))*tau
-             endif
+            if (associated(stexttau) ) then
+               where (ple(:,:,k) .le. tropp) 
+                  stexttau(:,:,w) = stexttau(:,:,w) + tau(:,:,k)
+               elsewhere(ple(:,:,k-1) .lt. tropp) 
+                 stexttau(:,:,w)  = stexttau(:,:,w) + log(tropp/ple(:,:,k-1))/log(ple(:,:,k)/ple(:,:,k-1))*tau(:,:,k)
+               endwhere
+            endif
+
+            if ( associated(scatau) ) then
+               scatau(:,:,w) = scatau(:,:,w) + tau(:,:,k)*ssa(:,:,k)
+            endif
+
+            if ( associated(stscatau) ) then
+               where (ple(:,:,k) .le. tropp) 
+                  stscatau(:,:,w) = stscatau(:,:,w) + tau(:,:,k)*ssa(:,:,k)
+               elsewhere(ple(:,:,k-1) .lt. tropp) 
+                  stscatau(:,:,w) = stscatau(:,:,w) + log(tropp/ple(:,:,k-1))/log(ple(:,:,k)/ple(:,:,k-1))*tau(:,:,k)*ssa(:,:,k)
+               endwhere
            endif
-
-           if( associated(scatau) ) then
-            scatau(i,j,w) = scatau(i,j,w) + tau*ssa
-           endif
-
-           if( associated(stscatau) ) then
-              if (ple(i,j,k) .le. tropp(i,j)) then
-                 stscatau(i,j,w) = stscatau(i,j,w) + tau*ssa
-              elseif(ple(i,j,k) .gt. tropp(i,j) .and. ple(i,j,k-1) .lt. tropp(i,j)) then
-                 stscatau(i,j,w) = stscatau(i,j,w) + log(tropp(i,j)/ple(i,j,k-1))/log(ple(i,j,k)/ple(i,j,k-1))*tau*ssa
-              endif
-           endif
-
-
          enddo
-        enddo
-       enddo
       enddo
    endif
 
@@ -7093,21 +6995,17 @@ K_LOOP: do k = km, 1, -1
       tau470(i1:i2,j1:j2) = tiny(1.0)
       tau870(i1:i2,j1:j2) = tiny(1.0)
 
-       do k = klid, km
-        do j = j1, j2
-         do i = i1, i2
+      call mie%Query(1, ilam470, & ! Only SO4 exists in the MieTable, so its index is 1
+             SO4*delp/grav, rh, tau=tau)
+      do k = klid, km
+         tau470 = tau470 + tau(:,:,k)
+      enddo
 
-          call Chem_MieQuery(mie_table, 1, ilam470, & ! Only SO4 exists in the MieTable, so its index is 1
-              SO4(i,j,k)*delp(i,j,k)/grav, rh(i,j,k), tau=tau)
-          tau470(i,j) = tau470(i,j) + tau
-
-          call Chem_MieQuery(mie_table, 1, ilam870, &
-              SO4(i,j,k)*delp(i,j,k)/grav,rh(i,j,k), tau=tau)
-          tau870(i,j) = tau870(i,j) + tau
-
-         enddo
-        enddo
-       enddo
+      call mie%Query(1, ilam870, &
+              SO4*delp/grav,rh, tau=tau)
+      do k = klid, km
+         tau870 = tau870 + tau(:,:,k)
+      enddo
 
 !      enddo  ! nbins
       angstrom(i1:i2,j1:j2) = &
@@ -7141,7 +7039,7 @@ K_LOOP: do k = km, 1, -1
         enddo
        endif
    endif
-
+   deallocate(tau,ssa)
    __RETURN__(__SUCCESS__)
    end subroutine SU_Compute_Diags